add exceptions + write LobbyController methods

+++ lobby-service/src/main/java/wonders/lobbyservice/service/RequestHandler.java
@@ -5,10 +5,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import wonders.lobbyservice.model.LobbyEntity;
 import wonders.lobbyservice.model.PlayerEntity;
+import wonders.lobbyservice.model.exception.NotFoundException;
 
 import java.sql.Time;
 import java.util.Collections;
 import java.util.HashMap;
+import java.util.Optional;
 
 @Service
 public class RequestHandler {
@@ -21,12 +23,14 @@ public class RequestHandler {
     /*
      * @return attributes to response
      */
-    public HashMap<String, String> createLobby(HashMap<String, String> attributes) throws Exception {
+    public HashMap<String, String> createLobby(HashMap<String, String> attributes) throws IllegalArgumentException {
         PlayerEntity player = new PlayerEntity();
 
         //TODO добавить начальный статус игрока, мб в enum
         if (attributes.containsKey("playerName")) {
             player.setUsername(String.valueOf(attributes.get("playerName")));
+        } else {
+            throw new IllegalArgumentException("missing player name attribute");
         }
 
         LobbyEntity lobby = new LobbyEntity();
@@ -37,17 +41,17 @@ public class RequestHandler {
         if (attributes.containsKey("lobbyName")) {
             lobby.setName(attributes.get("lobbyName"));
         } else {
-            throw new Exception();
+            throw new IllegalArgumentException("missing lobby name attribute");
         }
         if (attributes.containsKey("maxPlayers")) {
             lobby.setMaxPlayers(Integer.valueOf(attributes.get("maxPlayers")));
         } else {
-            throw new Exception();
+            throw new IllegalArgumentException("missing max player attribute");
         }
         if (attributes.containsKey("moveTime")) {
             lobby.setMoveTime(Time.valueOf(attributes.get("moveTime")));
         } else {
-            throw new Exception();
+            throw new IllegalArgumentException("missing move time attribute");
         }
 
         lobby = lobbyService.save(lobby);
@@ -67,11 +71,18 @@ public class RequestHandler {
      * @return attributes to response
      */
     @Transactional
-    public HashMap<String, String> deleteLobby (HashMap<String, String> attributes) throws Exception {
+    public HashMap<String, String> deleteLobby (HashMap<String, String> attributes) throws IllegalArgumentException {
         if(attributes.containsKey("lobbyId")){
+            Long lobbyId = (Long.valueOf(attributes.get("lobbyId")));
+            Optional<LobbyEntity> lobby = lobby = lobbyService.findById(lobbyId);
+
+            if(lobby.isEmpty()) {
+                throw new IllegalArgumentException(String.format("lobby with '%d' id doesn't exist", lobbyId));
+            }
+
             lobbyService.deleteById(Long.valueOf(attributes.get("lobbyId")));
         } else {
-            throw new Exception();
+            throw new IllegalArgumentException("missing lobby id attribute");
         }
 
         HashMap<String, String> results = new HashMap<>();
@@ -84,33 +95,35 @@ public class RequestHandler {
      * @return attributes to response
      */
     @Transactional
-    public HashMap<String, String> connectPlayer(HashMap<String, String> attributes) throws Exception {
+    public HashMap<String, String> connectPlayer(HashMap<String, String> attributes) throws IllegalArgumentException {
         PlayerEntity lobbyPlayer = new PlayerEntity();
-        LobbyEntity lobby;
+        Optional<LobbyEntity> lobby;
 
         if(attributes.containsKey("lobbyId")) {
             Long lobbyId = (Long.valueOf(attributes.get("lobbyId")));
 
-            if (lobbyService.existById(lobbyId)){
-                lobby = lobbyService.findById(lobbyId);
-            } else {
-                throw new Exception();
+            lobby = lobbyService.findById(lobbyId);
+
+            if(lobby.isEmpty()) {
+                throw new IllegalArgumentException(String.format("lobby with '%d' id doesn't exist", lobbyId));
             }
+
         } else {
-            throw new Exception();
+            throw new IllegalArgumentException("missing lobby id attribute");
         }
 
         if(attributes.containsKey("playerName")) {
             lobbyPlayer.setUsername(attributes.get("playerName"));
         } else {
-            throw new Exception();
+            throw new IllegalArgumentException("missing player name attribute");
         }
 
-        lobby.addPlayer(lobbyPlayer);
-        lobbyService.save(lobby);
+        LobbyEntity lobbyToUpdate = lobby.get();
+        lobbyToUpdate.addPlayer(lobbyPlayer);
+        lobbyService.save(lobbyToUpdate);
 
         HashMap<String, String> result = new HashMap<>();
-        result.put("lobbyId", lobby.getId().toString());
+        result.put("lobbyId", lobbyToUpdate.getId().toString());
         result.put("playerId", String.valueOf(lobbyPlayer.getId()));
         result.put("playerName", lobbyPlayer.getUsername());
